add endpoint Get para categorias e lancamentos

+++ src/main/java/br/com/FinanceApp/repository/LancamentoRepository.java
@@ -3,5 +3,5 @@ package br.com.FinanceApp.repository;
 import br.com.FinanceApp.entity.Lancamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LancamentoRepository extends JpaRepository<Lancamento, Integer> {
+public interface LancamentoRepository extends JpaRepository<Lancamento, Long> {
 }
